Make it possible to specify a custom clusterDomain if it differs from the default `cluster.local`
A new field `clusterDomain` is added and if not specified, it falls back to
the default of `cluster.local`.

Note that if you modify a `MinIOInstance`, the `StatefulSet` resource is _not_ modified currently in this commit.
So it's only taken into account when creating a `MinIOInstance`.

This should help for issue #70 and issue minio/minio#6775.