Remove user provided liveness and readiness config (#129)
This PR removes user provided configuration for liveness and
readiness probes, since most of these fields are expected to be
constant and are known to us, we should not require users to
provide these. Instead, we simply use all known values and create
the probes - we only expect initialDelay and probe period from
users. If these fields are not provided, we do not create the
probe.

Additionally, this PR also adds steps on how to create local PV
before creating a MinIOInstance.

Fixes #124