Improve the zone addition process (#150)
Currently the operator deletes existing statefulset and creates a
fresh new statefulset, during zone addition. This is not ideal
because this requires a downtime of MinIO cluster.

This PR attempts to improve this process by using the update
statefulset API to add new pods and then deletes only the pods
that were part of existing statefulset. As the pods are deleted
Kubernetes spawns new pods. This means at any point, more than
quorum number of MinIO pods are available. This allows a better,
seamless zone addition flow.

Also improve the minioinstance status update flow and cleanup
KES flow in controller.