Add zone to the Statefulset Match Labels (#238)
This adds the `ZoneLabel` to the Zone Statefulset matchLabels. The reasoning for adding it is because the PVCs created by the statefulset inherit the matchLabels of the Statefulset that created them. In order to properly asses which PVCs belong to a certain zone (without relying on the name of the PVC) we want to push down this label to the PVC from the statefulset.