Turn-off bucket DNS by default, make it opt-in
To enable one must enable this explicitly.
```
s3:
 bucketDNS: true
```

It looks like many S3 API users seem to have buckets with
`.` in them and would love to just use `path-style` instead
of `bucket` DNS style.

Making this an opt-in because users should be aware of the
restrictions this would bring in for their bucket names.