Do not set MINIO_SERVER_URL if already specified
MINIO_SERVER_URL is expected to get an address of the tenant accessible
from outside the kubernetes cluster.

The current code always adds this value by default and use the service
address, which is wrong when the user specifies this env variable in his
yaml.

This commit will only set an environment if not given by the user.