fix: Ignore audit config not found error when removing the log subsystem
The operator can be in a weird state when removing the log from the
tenant yaml. If removing the audit webhook succeed but the removal of
the log service fails, the operator will enter in an loop of errors
because it is not ignoring 'config not found' error when deleting
the webhook config again in the next try.

The commit ignores XMinioConfigError and XMinioConfigNotFoundError
though XMinioConfigError is not returned when a config is not found
since MinIO server Dec 2022.