Fix timing error leading to a false positive.
When the KES pod starts we read from the pod logs the "Root Token"
https://github.com/minio/operator/blob/5ef838a70a45d9a7de9a5400be153bcbdb2e5223/testing/console-tenant%2Bkes.sh#L73
However pod running doesn't mean vault process inside already reached that step on the init process.

Solution: We wait for the KES deploy Status instead of the pod running.

Then we wait for the log message to show, with timeout.

If the log message doesn't show, this time is a possitive fail.